Refactor the factories.
There is now an abstract factory class that ship as many common
methods as possible.
Let's achieve DRYness.
